refactor(Sprint6): 파일 이름을 kebab-case로 변경 (add-item-page)
